+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - [7950](https://github.com/vegaprotocol/vega/issues/7950) - Fix the restore of deposits from checkpoint
 - [7933](https://github.com/vegaprotocol/vega/issues/7933) - Ensure the wallet store is closed to avoid "too many opened files" error
 - [7956](https://github.com/vegaprotocol/vega/issues/7956) - Floor negative slippage per unit at 0
+- [7964](https://github.com/vegaprotocol/vega/issues/7964) - Use mark price for all margin calculations
+
 
 ## 0.70.0